+++ crates/ecs/src/generated.rs
@@ -1136,11 +1136,10 @@ pub mod messages {
     use glam::{Mat4, Quat, UVec2, UVec3, UVec4, Vec2, Vec3, Vec4};
     #[derive(Clone, Debug)]
     #[doc = "**Frame**: Sent to all modules every frame."]
-    pub struct Frame {}
+    pub struct Frame;
     impl Frame {
-        #[allow(clippy::too_many_arguments)]
         pub fn new() -> Self {
-            Self {}
+            Self
         }
     }
     impl Message for Frame {
@@ -1217,11 +1216,10 @@ pub mod messages {
     impl RuntimeMessage for ColliderLoads {}
     #[derive(Clone, Debug)]
     #[doc = "**ModuleLoad**: Sent to a module when it loads."]
-    pub struct ModuleLoad {}
+    pub struct ModuleLoad;
     impl ModuleLoad {
-        #[allow(clippy::too_many_arguments)]
         pub fn new() -> Self {
-            Self {}
+            Self
         }
     }
     impl Message for ModuleLoad {
@@ -1244,11 +1242,10 @@ pub mod messages {
     }
     #[derive(Clone, Debug)]
     #[doc = "**ModuleUnload**: Sent to a module when it unloads."]
-    pub struct ModuleUnload {}
+    pub struct ModuleUnload;
     impl ModuleUnload {
-        #[allow(clippy::too_many_arguments)]
         pub fn new() -> Self {
-            Self {}
+            Self
         }
     }
     impl Message for ModuleUnload {
@@ -1300,11 +1297,10 @@ pub mod messages {
     impl RuntimeMessage for WindowFocusChange {}
     #[derive(Clone, Debug)]
     #[doc = "**WindowClose**: Sent when the window is closed."]
-    pub struct WindowClose {}
+    pub struct WindowClose;
     impl WindowClose {
-        #[allow(clippy::too_many_arguments)]
         pub fn new() -> Self {
-            Self {}
+            Self
         }
     }
     impl Message for WindowClose {